TOC performance Bug:T91087 Bug:T91330
Instead of creating the TOC for the current article before it displays, use the DOM loaded notification and create it after all the web work is complete.
Only update the TOC in real time if visible.
Since we are only updating while the TOC is visible, update it on every scroll instead of when deceleration begins to keep the TOC in sync while scrolling
If hidden, update the TOC only after scrolling has ended.

Change-Id: Ib96275e9f7995b278b33cc56cad84da4678d5b29