first attempt at icon in fastlane
Big issue… after the app is built we only have the IPA available to us in fastlane - we need the bundle before it is turned into the IPA - which means we need to be in the build phases - which means we need to inject our logic into a build phase before the build starts.

For this task we are using the xcodeproj gem to inspect and modify the project file. As a first pass, just injecting the existing shell script. In the future it would be nice to rewrite the script in ruby.

Change-Id: I858c0af678b662a7db4d6711ebea1ca86c1c17c5